requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==2.0.23
google-cloud-secret-manager==2.17.0 # Or the latest stable version

